Day 13/1

+++ src/main/java/com/example/JobHunter/controller/admin/UserController.java
@@ -1,17 +1,27 @@
 package com.example.JobHunter.controller.admin;
 
+import com.example.JobHunter.Util.annotation.ApiMessage;
+import com.example.JobHunter.domain.RestResponse;
 import com.example.JobHunter.domain.User;
 import com.example.JobHunter.domain.dto.ResultPaginationDTO;
+import com.example.JobHunter.domain.dto.UserCreateDTO;
+import com.example.JobHunter.domain.dto.UserDTO;
+import com.example.JobHunter.domain.dto.UserUpdateDTO;
 import com.example.JobHunter.service.UserService;
 import com.turkraft.springfilter.boot.Filter;
+
+import jakarta.validation.Valid;
+
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
+@RequestMapping("/api/v1")
 public class UserController {
 
     private UserService userService;
@@ -22,14 +32,22 @@ public class UserController {
         this.passwordEncoder = passwordEncoder;
     }
 
-    @PostMapping("/users/create")
-    public ResponseEntity<User> create(@RequestBody User user) {
+    @PostMapping("/users")
+    @ApiMessage("Create a new user")
+    public ResponseEntity<UserCreateDTO> create(@Valid @RequestBody User user) {
         // return this.userService.createUser(user);
+        if (this.userService.existsByEmail(user.getEmail())) {
+            throw new UsernameNotFoundException("Email " + user.getEmail() + " already exists");
+        }
+
         user.setPassword(passwordEncoder.encode(user.getPassword()));
-        return ResponseEntity.status(HttpStatus.CREATED).body(this.userService.createUser(user));
+        User userOld = this.userService.createUser(user);
+
+        return ResponseEntity.status(HttpStatus.CREATED).body(this.userService.convertToUserCreateDTO(userOld));
     }
 
     @GetMapping("/users")
+    @ApiMessage("Fetch all users")
     public ResponseEntity<ResultPaginationDTO> getUsers(
             @Filter Specification<User> spec, Pageable pageable) {
 
@@ -37,17 +55,35 @@ public class UserController {
     }
 
     @GetMapping("/users/{id}")
-    public ResponseEntity<User> getUser(@PathVariable long id) {
-        return ResponseEntity.ok(this.userService.getUserById(id));
+    @ApiMessage("Fetch user by id")
+    public ResponseEntity<UserDTO> getUser(@PathVariable("id") long id) {
+        if (this.userService.fetchUserByid(id) == null) {
+            throw new UsernameNotFoundException("Not found user with id: " + id);
+        }
+        User user = this.userService.getUserById(id);
+        return ResponseEntity.ok(this.userService.convertToUserDTO(user));
     }
 
-    @PutMapping("/users/update")
-    public ResponseEntity<User> updateUser(@RequestBody User user) {
-        return ResponseEntity.ok(this.userService.updateUser(user));
+    @PutMapping("/users")
+    @ApiMessage("Update user")
+    public ResponseEntity<UserUpdateDTO> updateUser(@RequestBody User user) {
+        if (user != null) {
+            if (this.userService.fetchUserByid(user.getId()) == null) {
+                throw new UsernameNotFoundException("Not found user with id: " + user.getId());
+            }
+        }
+        User userOld = this.userService.updateUser(user);
+        return ResponseEntity.ok(this.userService.convertToUserUpdateDTO(userOld));
     }
 
     @DeleteMapping("/users/delete/{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable long id) {
+    @ApiMessage("Delete user by id")
+    public ResponseEntity<Void> deleteUser(@PathVariable("id") long id) {
+
+        if (this.userService.fetchUserByid(id) == null) {
+            throw new UsernameNotFoundException("Not found user with id: " + id);
+
+        }
         return ResponseEntity.ok(this.userService.deleteUserByid(id));
     }
